cria comunicacao com todos os servicos em relacao a realizar pedido

+++ email/src/main/java/br/com/emanueldias/email/dto/InfoPedidoDTO.java
@@ -5,6 +5,6 @@ import java.util.List;
 
 public record InfoPedidoDTO(
         List<String> itens,
-        BigDecimal valor
+        Double valor
 ) {
 }
